pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlalchemy = "^2.0.32"
aiomysql = "^0.2.0"
alembic = "^1.13.2"
mysqlclient = "^2.2.4"
pymysql = "^1.1.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.8"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
